move LazyValue to collection package
