+++ src/main/kotlin/gregtechlite/gtlitecore/api/collection/LazyValue.kt
@@ -1,8 +1,7 @@
-package gregtechlite.gtlitecore.api.block.attribute
+package gregtechlite.gtlitecore.api.collection
 
 class LazyValue<T>(private val delegate: () -> T)
 {
-
     @Volatile
     private var cachedValue: T? = null
 
@@ -14,5 +13,4 @@ class LazyValue<T>(private val delegate: () -> T)
     fun get() = cachedValue ?: synchronized(this) {
         cachedValue ?: delegate().also { cachedValue = it }
     }
-
 }
